Optimize triangle level algoritm
After some research I found a simple algorithm that can generate an arbitrary level without
the need for the previous level. This is very useful since one of the bottlenecks for the
previous implementation was keeping the "previous level" in memory and in my rough
benchmark this increases performance after about 4500 levels, but more importantly reduces
the memory bottleneck. This could be further optimized to check inline  the odd-ness, but
I'm skeptical that the potential performance improvement may be outweighed by the obfuscation
of the code, as the count and total values would need to be passed around a bit more.